add ability to set image size in feature_row
This follows the guidance in https://github.com/mmistakes/minimal-mistakes/issues/2187
with the addition of an if/endif to handle the case where image_size is not
specified.

To use, provide the `image_size: value` attribute for an element
in a feature row. The value can be a value (in pixels) or a
percentage (of the bounding box for the feature).